Revert "Move unified-latex to worker pool"
This reverts commit 75c4e36f57a69bd9f487d5c3c349ccc5774da345.
