+++ src/commander.ts
@@ -449,7 +449,7 @@ export async function devParseTeX() {
     if (vscode.window.activeTextEditor === undefined) {
         return
     }
-    const ast = await parser.unifiedParse(vscode.window.activeTextEditor.document.getText())
+    const ast = parser.unifiedParse(vscode.window.activeTextEditor.document.getText())
     return vscode.workspace.openTextDocument({content: JSON.stringify(ast, null, 2), language: 'json'}).then(doc => vscode.window.showTextDocument(doc))
 }
