+++ src/providers/structurelib/doctex.ts
@@ -46,7 +46,7 @@ function getDoc(content: string) {
 }
 
 async function getToC(document: vscode.TextDocument, docContent: string) {
-    const ast = await parser.unifiedParse(docContent)
+    const ast = parser.unifiedParse(docContent)
     if (ast === undefined) {
         logger.log('Failed parsing LaTeX AST.')
         return []
